pyproject.toml
[tool.poetry]
name = "spotiseek"
version = "0.1.0"
description = ""
authors = ["KijkEr <joep_is_cool@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
slskd-api = "^0.1.5"
spotipy = "^2.24.0"
thefuzz = "^0.22.1"
typer = "^0.12.5"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
spotiseek = "spotiseek.cli:app"
